tests/conftest.py
import pytest, sys, random, py, pytest, os
from xprocess import ProcessStarter
from app import app

# This fixture starts the test server and makes it available to the tests.
# You don't need to understand it in detail.
@pytest.fixture
def test_web_address(xprocess):
    python_executable = sys.executable
    app_file = py.path.local(__file__).dirpath("../app.py")
    port = "5001"
    class Starter(ProcessStarter):
        env = {"PORT": port, "APP_ENV": "test", **os.environ}
        pattern = "Debugger PIN"
        args = [python_executable, app_file]

    xprocess.ensure("flask_test_server", Starter)

    yield f"127.0.0.1:{port}"

    xprocess.getinfo("flask_test_server").terminate()
