requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
playwright==1.42.0
pluggy==1.4.0
psutil==5.9.8
pyee==11.0.1
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
pytest-xprocess==0.23.0
python-slugify==8.0.4
requests==2.31.0
text-unidecode==1.3
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
